[REFACTOR] #20: 시간당 가격 컬럼 타입 수정 (->Integer)

+++ src/main/java/project/luckybooky/domain/location/entity/Location.java
@@ -30,7 +30,7 @@ public class Location extends BaseEntity {
     private String locationImageUrl;
 
     @Column(name = "price_per_hour", nullable = false)
-    private Long pricePerHour;
+    private Integer pricePerHour;
 
     @Column(name = "seat_count", nullable = false)
     private Integer seatCount;
